Allow to install from a git branch
The tricky part is that pip is a bit dumb since a fiew version and
would not reinstall the package even from new source/new branch/new
reference if the package version itself does not change. So it seems
we are forced to use `--force-reinstall`, which is silly and kills
trees.